Move Material Table helper out of ProjectTeamTable, finish editable actions obj
